+++ index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en" class=" ">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <meta name="keywords" content="Motorcycle, bikers, biker, motorcycle rider, motorcycle clubs, social media, social bikers, biker social, motorcycle repair, group rides, share routes, discover routes">
+  <title>Moto Meetup - Where Bikers Connect</title>
+  <link rel="stylesheet" href="styles.css">
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.min.css">
+  <link href="https://fonts.googleapis.com/css?family=Merriweather:400,900,900i|Montserrat&display=swap" rel="stylesheet">
+  <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.12.1/css/all.min.css" rel="stylesheet">
+</head>
+<body id="desktop-template" class="unifiedNav" style>
+
+
+  <footer>Copyright 2022</footer>
+  <div id="root"></div>
+  <script type="module" src="index.js"></script>
+</body>
+  </html>
