+++ maps.html
@@ -10,8 +10,8 @@
     <link href="https://fonts.googleapis.com/css?family=Merriweather:400,900,900i|Montserrat&display=swap" rel="stylesheet">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.12.1/css/all.min.css" rel="stylesheet">
   </head>
-  
-    <body> 
+
+    <body>
   <header>
     <h1>MOTOMEETUP</h1>
   </header>
@@ -19,9 +19,9 @@
     <i class="fas fa-bars"></i>
     <ul class="hidden--mobile nav-links">
       <li><a href="index.html">Home</a></li>
-      <li><a href="about.html">Bio</a></li>
-      <li><a href= "maps.html">Pizza</a></li>
-      <li><a href= "groups.html">Order</a></li>
+      <li><a href="about.html">About</a></li>
+      <li><a href= "maps.html">Maps</a></li>
+      <li><a href= "groups.html">Groups</a></li>
     </ul>
   </nav>
   <section id="maps">
@@ -37,29 +37,13 @@
         </select>
       </div>
       <h2>Let's plan a route</h2>
-<div> 
+<div>
 <label for="Location">Map:</label>
         <select id="Local" name="Rides">
-          <option value="">'GET'</option>
+          <option value="">Let's go for a Ride</option>
             </select>
       </div>
-const axios = require("axios");
 
-const options = {
-  method: 'GET',
-  url: 'https://google-maps-geocoding.p.rapidapi.com/geocode/json',
-  params: {latlng: '40.714224,-73.96145', language: 'en'},
-  headers: {
-    'X-RapidAPI-Key': '6abfe24913msh71dbee813a74c24p1e6de0jsn9e02cda21d8a',
-    'X-RapidAPI-Host': 'google-maps-geocoding.p.rapidapi.com'
-  }
-};
-
-axios.request(options).then(function (response) {
-	console.log(response.data);
-}).catch(function (error) {
-	console.error(error);
-});
     &copy; 2020 <a href="https://savvycoders.com/">Savvy Coders</a>
   </footer>
   <script src="index.js"></script>
