+++ package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "scripts": {},
+  "scripts": { "parcel-build": "parcel build index.html && cp _redirects ./dist/",
+  "serve": "parcel index.html"},
   "keywords": [],
   "author": "",
   "license": "ISC",
@@ -20,8 +21,4 @@
     "yargs-parser": ">=13.1.2"
   }
 }
-"scripts": {
-    "parcel-build": "parcel build index.html && cp _redirects ./dist/",
-    "serve": "parcel index.html"
-  },
-...
+
